+++ jest.config.json
@@ -13,3 +13,4 @@
   ],
   "testRegex": "src\/.*\\.test\\.jsx?$"
 }
+
